refactor tx hooks folder
