+++ src/pages/Transactions/components/TxDetail/index.tsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux'
 import { getExplorerInfo } from 'src/config'
 import { grantedSelector } from 'src/routes/safe/container/selector'
 import { formatWithSchema } from 'src/utils/date'
-import { useTransactionDetails } from '../../hooks/useTransactionDetails'
+import { useTransactionDetails } from '../../../../utils/hooks/useTransactionDetails'
 import { Centered, InlineEthHashInfo, TxDetailsContainer } from '../../styled'
 import { TxActions } from './Action'
 import TxMsg from './Message'
