+++ src/pages/Voting/index.tsx
@@ -22,7 +22,7 @@ import { getProposals, MChainInfo } from 'src/services'
 import { IProposal } from 'src/types/proposal'
 import { calcBalance } from 'src/utils/calc'
 import { formatDateTimeDivider } from 'src/utils/date'
-import { usePagedQueuedTransactions } from '../Transactions/hooks/usePagedQueuedTransactions'
+import { usePagedQueuedTransactions } from '../../utils/hooks/usePagedQueuedTransactions'
 import ProposalsCard from './ProposalsCard'
 import { ProposalsSection, StyledBlock, StyledColumn, TitleNumberStyled } from './styledComponents'
 import VotingModal from './VotingPopup'
